requirements.txt
numpy
torch
torchaudio
librosa
pydub
transformers
pyannote.audio
speechbrain>=1.0.0
ffmpeg
git+https://github.com/openai/whisper.git
